+++ src/lib.rs
@@ -9,7 +9,7 @@
 //!   to the well-defined directory.
 //! - Using methods under [`mod@flag`] to use thread-safe, lock-free flags.
 //! - Using methods under [`mod@slack`] to send notifications to Slack with Slack Hook.
-//! - Using methods under [`mod@message`] to subscribe and publish based on Nanomsg.
+//! - Using methods under [`mod@message`] to subscribe and publish based on Nanomsg or Zeromq.
 //!
 //! These utilities are especial for [wmjtyd/crypto-market](https://github.com/wmjtyd/crypto-market);
 //! however, you can also use it in your project. We licensed it under `Apache-2.0`, the same
@@ -26,5 +26,4 @@ pub mod flag;
 #[cfg(feature = "slack")]
 pub mod slack;
 
-#[cfg(feature = "message")]
 pub mod message;